--- src/main/java/practice/buttersaltflour/domain/member/entity/Gender.java
@@ -1,6 +1,0 @@
-package practice.buttersaltflour.domain.member.entity;
-
-public enum Gender {
-    MALE,
-    FEMALE
-}
