+++ src/main/java/practice/buttersaltflour/domain/member/entity/Youth.java
@@ -5,6 +5,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import practice.buttersaltflour.domain.member.entity.enumList.Gender;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
