+++ src/main/java/practice/buttersaltflour/domain/member/entity/enumList/Gender.java
@@ -1,0 +1,5 @@
+package practice.buttersaltflour.domain.member.entity.enumList;
+
+public enum Gender {
+    FEMALE, MALE
+}
