+++ src/main/java/practice/buttersaltflour/domain/member/entity/enumList/MatchStatus.java
@@ -1,0 +1,5 @@
+package practice.buttersaltflour.domain.member.entity.enumList;
+
+public enum MatchStatus {
+    PENDING, APRROVED, COMPLETED
+}
