--- src/main/java/practice/buttersaltflour/member/entity/enumList/Gender.java
@@ -1,5 +1,0 @@
-package practice.buttersaltflour.member.entity.enumList;
-
-public enum Gender {
-    FEMALE, MALE
-}
