--- src/main/java/practice/buttersaltflour/member/entity/enumList/MatchStatus.java
@@ -1,5 +1,0 @@
-package practice.buttersaltflour.member.entity.enumList;
-
-public enum MatchStatus {
-    PENDING, APRROVED, COMPLETED
-}
